# apps/estimate-room-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WebSocket-based room estimation service.",
        "title": "Estimate Room API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/auth/forgot-password": {
            "post": {
                "description": "Accepts an email and stores a one-time reset token when the account is eligible. The response is always generic.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start password reset",
                "parameters": [
                    {
                        "description": "Email address",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.ForgotPasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.ForgotPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/github/callback": {
            "get": {
                "description": "Completes GitHub login, creates a browser session, and redirects back to the OAuth continue URL.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "GitHub auth callback",
                "responses": {
                    "302": {
                        "description": "Redirect to continue URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/github/login": {
            "get": {
                "description": "Redirects the browser to GitHub OAuth using the provided OAuth continue URL.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start GitHub login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth continue URL",
                        "name": "continue",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to GitHub",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates a local user, creates a browser session, and returns the authenticated user payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with email and password",
                "parameters": [
                    {
                        "description": "Credentials and OAuth continue URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Revokes the current browser session and clears auth cookies.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.LogoutResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Creates a local user, auto-signs them in, and returns the authenticated user payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a local account",
                "parameters": [
                    {
                        "description": "Registration payload and OAuth continue URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.RegisterDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/authdto.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "description": "Resets the password for a valid token and revokes all active sessions and tokens for that user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset local account password",
                "parameters": [
                    {
                        "description": "Reset token and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.ResetPasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.ResetPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/reset-password/validate": {
            "get": {
                "description": "Returns whether a password reset token is valid, expired, invalid, or already used.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate password reset token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reset token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.ResetPasswordValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/session": {
            "get": {
                "description": "Returns the currently authenticated user for the active browser session, or authenticated=false when there is no active session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current browser auth session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.SessionResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorItem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "pointer": {
                    "type": "string"
                }
            }
        },
        "apperrors.HttpError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apperrors.ErrorItem"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "authdto.ForgotPasswordDTO": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "authdto.ForgotPasswordResponse": {
            "type": "object",
            "properties": {
                "submitted": {
                    "type": "boolean"
                }
            }
        },
        "authdto.LoginDTO": {
            "type": "object",
            "required": [
                "continue",
                "email",
                "password"
            ],
            "properties": {
                "continue": {
                    "type": "string",
                    "maxLength": 2000
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
        "authdto.LogoutResponse": {
            "type": "object",
            "properties": {
                "loggedOut": {
                    "type": "boolean"
                }
            }
        },
        "authdto.RegisterDTO": {
            "type": "object",
            "required": [
                "continue",
                "email",
                "password"
            ],
            "properties": {
                "continue": {
                    "type": "string",
                    "maxLength": 2000
                },
                "displayName": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "occupation": {
                    "type": "string",
                    "maxLength": 120
                },
                "organization": {
                    "type": "string",
                    "maxLength": 120
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
        "authdto.ResetPasswordDTO": {
            "type": "object",
            "required": [
                "password",
                "token"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                },
                "token": {
                    "type": "string",
                    "maxLength": 512
                }
            }
        },
        "authdto.ResetPasswordResponse": {
            "type": "object",
            "properties": {
                "reset": {
                    "type": "boolean"
                }
            }
        },
        "authdto.ResetPasswordValidationResponse": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "authdto.SessionResponse": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/authdto.SessionUserResponse"
                }
            }
        },
        "authdto.SessionUserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                }
            }
        },
...
    }
}